Split homebrew initialization
The homebrew env vars need setup during profile init so that the env
vars are exported appropriately.

However, the sourcing of completion scripts needs to occur during bashrc
init, so that completion is available to all shells and subshells.

So introduce a new pattern whereby .config/bashrc.d/*.sh will be sourced
by .profile, and *.bash files will be sourced by .bashrc.

This way we can minimize the amount of work that gets done on each shell
instantiation.

For clarity:
.profile init occurs for new Terminal windows on macOS
.bashrc init occurs for subshells and new Terminal _tabs_